teco_string_get_coord() returns character offsets now (refs #5)
* This is used for error messages (TECO macro stackframes),
  so it's important to display columns in characters.
* Program counters are in bytes and therefore everywhere gsize.
  This is by glib convention.

+++ src/cmdline.h
@@ -46,7 +46,7 @@ typedef struct {
 	gsize effective_len;
 
 	/** Program counter within the command-line macro */
-	guint pc;
+	gsize pc;
 
 	/**
 	 * Specifies whether the immediate editing modifier
